use int instead of string for device- and groupIds
